Add pipeline finalizer
The pipeline lacked a feature where the last result isn't interceptable.
This is mainly used in nested pipelines (e.g. parallel)
