+++ result.go
@@ -1,0 +1,11 @@
+package pipeline
+
+// IsSuccessful returns true if the contained error is nil.
+func (r Result) IsSuccessful() bool {
+	return r.Err == nil
+}
+
+// IsFailed returns true if the contained error is non-nil.
+func (r Result) IsFailed() bool {
+	return r.Err != nil
+}
